+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Target Scheduler
 
-## 5.8.1.1 (beta) - 2025-08-07
-* Fix for threading problem in scheduler progress vm add
+## 5.8.2.0 - 2025-XX-XX
+* New profile preference to disable target completion reset at the profile level.
+* Fix for threading problem in scheduler progress vm add.
 
 ## 5.8.1.0 (beta) - 2025-08-07
 * Fixed issue with layout during exposure template editing.